% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraseña.R
\name{generar_clave}
\alias{generar_clave}
\title{Generador de nombre de usuario y contraseña segura}
\usage{
generar_clave(codigo, nombre = "silvina", apellido = "perez")
}
\arguments{
\item{codigo}{Entero de dos cifras. Número que se usará en el cálculo de la contraseña.}

\item{nombre}{Caracter. Nombre del usuario. Por defecto es \code{"silvina"}.}

\item{apellido}{Caracter. Apellido del usuario. Por defecto es \code{"perez"}.}
}
\value{
Una lista con dos elementos:
\describe{
\item{nombre_usuario}{El nombre de usuario en formato \code{nombre_apellido}, todo en minúsculas.}
\item{contraseña}{Una cadena que conforma una contraseña creada con los datos aportados}
}
}
\description{
Esta función genera un nombre de usuario y una contraseña combinando partes
del nombre, apellido y un número, incluyendo un número aleatorio.
}
\details{
La contraseña se construye de la siguiente manera:
\itemize{
\item 3 primeras letras del apellido (minúsculas)
\item cantidad de vocales del nombre multiplicada por la cantidad de consonantes del apellido
\item 2 primeras letras del nombre
\item código multiplicado por un número aleatorio del 1 al 10
}

Se imprime un mensaje indicando que la contraseña fue generada correctamente.
}
\examples{
generar_clave(42, "Juan", "Martinez")
generar_clave(32)

}
